pyproject.toml
[tool.poetry]
name = "에어비엔비 클론코딩"
version = "0.1.0"
description = ""
authors = ["Sungin Lee <dlrkehrud@naver.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.6"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-environ = "^0.10.0"
django-cors-headers = "^3.14.0"
requests = "^2.28.2"


[tool.poetry.group.dev.dependencies]
black = {version = "^23.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
